✨ Feat: manage dashboard list with zustand
